server/requirements.txt
Jinja2
MarkupSafe
PyJWT
PyYAML
Pygments
SQLAlchemy
annotated-types
anyio
asyncpg
certifi
click
dnspython
email_validator
fastalchemy
fastapi
fastapi-cli
fastapi-login
greenlet
h11
httpcore
httptools
httpx
idna
librosa
markdown-it-py
mdurl
pandas
passlib
psycopg2-binary
pydantic
pydantic_core
python-dotenv
python-multipart
rich
rich-toolkit
shellingham
sniffio
sqlalchemy
starlette
torch
torchaudio
torchvision
typer
typing_extensions
uvicorn[standard]
uvloop
watchfiles
websockets

transformers
phonemizer
pydub
